Factoring transaction and block types out of the protocol folder
