+++ Network/Haskoin/Protocol/Message.hs
@@ -27,10 +27,12 @@ import qualified Data.ByteString as BS
     )
 
 import Network.Haskoin.Protocol.Types
-
-import Network.Haskoin.Util 
+import Network.Haskoin.Transaction.Types
+import Network.Haskoin.Block.Types
 import Network.Haskoin.Crypto.Hash 
+import Network.Haskoin.Crypto.Merkle
 import Network.Haskoin.Constants
+import Network.Haskoin.Util 
 
 -- | Data type representing the header of a 'Message'. All messages sent between
 -- nodes contain a message header.
