+++ Network/Haskoin/Protocol/Types.hs
@@ -2,22 +2,14 @@ module Network.Haskoin.Protocol.Types
 ( Addr(..)
 , NetworkAddressTime 
 , Alert(..)
-, Block(..)
-, BlockHeader(..) 
 , BloomFlags(..)
 , BloomFilter(..)
 , FilterLoad(..)
 , FilterAdd(..)
-, GetBlocks(..) 
-, BlockLocator
 , GetData(..)
-, GetHeaders(..)
-, Headers(..)
-, BlockHeaderCount
 , Inv(..)
 , InvVector(..) 
 , InvType(..)
-, MerkleBlock(..)
 , NetworkAddress(..)
 , NotFound(..)
 , Ping(..)
@@ -25,11 +17,6 @@ module Network.Haskoin.Protocol.Types
 , Reject(..)
 , RejectCode(..)
 , reject
-, Tx(..) 
-, TxIn(..)
-, TxOut(..)
-, OutPoint(..)
-, CoinbaseTx(..)
 , VarInt(..)
 , VarString(..)
 , Version(..)
@@ -124,85 +111,6 @@ instance Binary Alert where
     get = Alert <$> get <*> get
     put (Alert p s) = put p >> put s
 
--- | Data type describing a block in the bitcoin protocol. Blocks are sent in
--- response to 'GetData' messages that are requesting information from a
--- block hash.
-data Block = 
-    Block {
-            -- | Header information for this block.
-            blockHeader     :: !BlockHeader
-            -- | Coinbase transaction of this block.
-          , blockCoinbaseTx :: !CoinbaseTx
-            -- | List of transactions pertaining to this block.
-          , blockTxns       :: ![Tx]
-          } deriving (Eq, Show, Read)
-
-instance NFData Block where
-    rnf (Block h c ts) = rnf h `seq` rnf c `seq` rnf ts
-
-instance Binary Block where
-
-    get = do
-        header     <- get
-        (VarInt c) <- get
-        cb         <- get
-        txs        <- replicateM (fromIntegral (c-1)) get
-        return $ Block header cb txs
-
-    put (Block h cb txs) = do
-        put h
-        put $ VarInt $ fromIntegral $ (length txs) + 1
-        put cb
-        forM_ txs put
-
--- | Data type recording information on a 'Block'. The hash of a block is
--- defined as the hash of this data structure. The block mining process
--- involves finding a partial hash collision by varying the nonce in the
--- 'BlockHeader' and/or additional randomness in the 'CoinbaseTx' of this
--- 'Block'. Variations in the 'CoinbaseTx' will result in different merkle 
--- roots in the 'BlockHeader'.
-data BlockHeader = 
-    BlockHeader {
-                  -- | Block version information, based on the version of the
-                  -- software creating this block.
-                  blockVersion   :: !Word32
-                  -- | Hash of the previous block (parent) referenced by this
-                  -- block.
-                , prevBlock      :: !BlockHash
-                  -- | Root of the merkle tree of all transactions pertaining
-                  -- to this block.
-                , merkleRoot     :: !Word256
-                  -- | Unix timestamp recording when this block was created
-                , blockTimestamp :: !Word32
-                  -- | The difficulty target being used for this block
-                , blockBits      :: !Word32
-                  -- | A random nonce used to generate this block. Additional
-                  -- randomness is included in the coinbase transaction of
-                  -- this block.
-                , bhNonce        :: !Word32
-                } deriving (Eq, Show, Read)
-
-instance NFData BlockHeader where
-    rnf (BlockHeader v p m t b n) =
-        rnf v `seq` rnf p `seq` rnf m `seq` rnf t `seq` rnf b `seq` rnf n
-
-instance Binary BlockHeader where
-
-    get = BlockHeader <$> getWord32le
-                      <*> get
-                      <*> get
-                      <*> getWord32le
-                      <*> getWord32le
-                      <*> getWord32le
-
-    put (BlockHeader v p m bt bb n) = do
-        putWord32le v
-        put         p
-        put         m
-        putWord32le bt
-        putWord32le bb
-        putWord32le n 
-
 -- | The bloom flags are used to tell the remote peer how to auto-update
 -- the provided bloom filter. 
 data BloomFlags
@@ -294,46 +202,6 @@ instance Binary FilterAdd where
         put $ VarInt $ fromIntegral $ BS.length bs
         putByteString bs
 
-type BlockLocator = [BlockHash]
-
--- | Data type representing a GetBlocks message request. It is used in the
--- bitcoin protocol to retrieve blocks from a peer by providing it a
--- 'BlockLocator' object. The 'BlockLocator' is a sparse list of block hashes
--- from the caller node with the purpose of informing the receiving node
--- about the state of the caller's blockchain. The receiver node will detect
--- a wrong branch in the caller's main chain and send the caller appropriate 
--- 'Blocks'. The response to a 'GetBlocks' message is an 'Inv' message
--- containing the list of block hashes pertaining to the request. 
-data GetBlocks = 
-    GetBlocks {
-                -- | The protocol version
-                getBlocksVersion  :: !Word32
-                -- | Block locator object. It is a list of block hashes from the
-                -- most recent block back to the genesis block. The list is
-                -- dense at first and sparse towards the end.
-              , getBlocksLocator  :: !BlockLocator
-                -- | Hash of the last desired block. If set to zero, the
-                -- maximum number of block hashes is returned (500).
-              , getBlocksHashStop :: !BlockHash
-              } deriving (Eq, Show, Read)
-
-instance NFData GetBlocks where
-    rnf (GetBlocks v l h) = rnf v `seq` rnf l `seq` rnf h
-
-instance Binary GetBlocks where
-
-    get = GetBlocks <$> getWord32le
-                    <*> (repList =<< get)
-                    <*> get
-      where 
-        repList (VarInt c) = replicateM (fromIntegral c) get
-
-    put (GetBlocks v xs h) = do
-        putWord32le v
-        put $ VarInt $ fromIntegral $ length xs
-        forM_ xs put
-        put h
-
 -- | The 'GetData' type is used to retrieve information on a specific object
 -- ('Block' or 'Tx') identified by the objects hash. The payload of a 'GetData'
 -- request is a list of 'InvVector' which represent all the hashes for which a
@@ -360,67 +228,6 @@ instance Binary GetData where
     put (GetData xs) = do
         put $ VarInt $ fromIntegral $ length xs
         forM_ xs put
-
--- | Similar to the 'GetBlocks' message type but for retrieving block headers
--- only. The response to a 'GetHeaders' request is a 'Headers' message
--- containing a list of block headers pertaining to the request. A maximum of
--- 2000 block headers can be returned. 'GetHeaders' is used by thin (SPV)
--- clients to exclude block contents when synchronizing the blockchain.
-data GetHeaders = 
-    GetHeaders {
-                 -- | The protocol version
-                 getHeadersVersion  :: !Word32
-                 -- | Block locator object. It is a list of block hashes from
-                 -- the most recent block back to the Genesis block. The list
-                 -- is dense at first and sparse towards the end.
-               , getHeadersBL       :: !BlockLocator
-                 -- | Hash of the last desired block header. When set to zero,
-                 -- the maximum number of block headers is returned (2000)
-               , getHeadersHashStop :: !BlockHash
-               } deriving (Eq, Show, Read)
-
-instance NFData GetHeaders where
-    rnf (GetHeaders v l h) = rnf v `seq` rnf l `seq` rnf h
-
-instance Binary GetHeaders where
-
-    get = GetHeaders <$> getWord32le
-                     <*> (repList =<< get)
-                     <*> get
-      where 
-        repList (VarInt c) = replicateM (fromIntegral c) get
-
-    put (GetHeaders v xs h) = do
-        putWord32le v
-        put $ VarInt $ fromIntegral $ length xs
-        forM_ xs put
-        put h
-
--- | 'BlockHeader' type with a transaction count as 'VarInt'
-type BlockHeaderCount = (BlockHeader, VarInt)
-
--- | The 'Headers' type is used to return a list of block headers in
--- response to a 'GetHeaders' message.
-data Headers = 
-    Headers { 
-              -- | List of block headers with respective transaction counts
-              headersList :: ![BlockHeaderCount] 
-            } 
-    deriving (Eq, Show, Read)
-
-instance NFData Headers where
-    rnf (Headers l) = rnf l
-
-instance Binary Headers where
-
-    get = Headers <$> (repList =<< get)
-      where 
-        repList (VarInt c) = replicateM (fromIntegral c) action
-        action = liftM2 (,) get get
-
-    put (Headers xs) = do
-        put $ VarInt $ fromIntegral $ length xs
-        forM_ xs $ \(a,b) -> put a >> put b
 
 -- | 'Inv' messages are used by nodes to advertise their knowledge of new
 -- objects by publishing a list of hashes. 'Inv' messages can be sent
@@ -488,62 +295,6 @@ instance NFData InvVector where
 instance Binary InvVector where
     get = InvVector <$> get <*> get
     put (InvVector t h) = put t >> put h
-
-data MerkleBlock = 
-    MerkleBlock {
-                -- | Header information for this merkle block.
-                  merkleHeader :: !BlockHeader
-                -- | Number of transactions in the block (including
-                -- unmatched transactions).
-                , merkleTotalTxns :: !Word32
-                  -- | Hashes in depth-first order. They are used to rebuild a
-                  -- partial merkle tree.
-                , mHashes :: ![Word256]
-                  -- | Flag bits, packed per 8 in a byte. Least significant bit
-                  -- first. Flag bits are used to rebuild a partial merkle
-                  -- tree.
-                , mFlags :: ![Bool]
-                } deriving (Eq, Show, Read)
-
-instance NFData MerkleBlock where
-    rnf (MerkleBlock m t h f) = rnf m `seq` rnf t `seq` rnf h `seq` rnf f
-
-instance Binary MerkleBlock where
-
-    get = do
-        header <- get
-        ntx    <- getWord32le
-        (VarInt matchLen) <- get
-        hashes <- replicateM (fromIntegral matchLen) get
-        (VarInt flagLen)  <- get
-        ws <- replicateM (fromIntegral flagLen) getWord8
-        return $ MerkleBlock header ntx hashes (decodeMerkleFlags ws)
-
-    put (MerkleBlock h ntx hashes flags) = do
-        put h
-        putWord32le ntx
-        put $ VarInt $ fromIntegral $ length hashes
-        forM_ hashes put
-        let ws = encodeMerkleFlags flags
-        put $ VarInt $ fromIntegral $ length ws
-        forM_ ws putWord8
-
-decodeMerkleFlags :: [Word8] -> [Bool]
-decodeMerkleFlags ws = 
-    [ b | p <- [0..(length ws)*8-1]
-    , b <- [testBit (ws !! (p `div` 8)) (p `mod` 8)]
-    ]
-
-encodeMerkleFlags :: [Bool] -> [Word8]
-encodeMerkleFlags bs = map boolsToWord8 $ splitIn 8 bs
-    
-splitIn :: Int -> [a] -> [[a]]
-splitIn _ [] = []
-splitIn c xs = take c xs : (splitIn c $ drop c xs)
- 
-boolsToWord8 :: [Bool] -> Word8
-boolsToWord8 [] = 0
-boolsToWord8 xs = foldl setBit 0 (map snd $ filter fst $ zip xs [0..7])
 
 -- | Data type describing a bitcoin network address. Addresses are stored in
 -- IPv6. IPv4 addresses are mapped to IPv6 using IPv4 mapped IPv6 addresses:
@@ -715,189 +466,6 @@ instance Binary Reject where
         put $ VarString $ stringToBS $ commandToString cmd
         put code
         put reason
-
--- | Data type representing a bitcoin transaction
-data Tx = 
-    Tx { 
-         -- | Transaction data format version
-         txVersion  :: !Word32
-         -- | List of transaction inputs
-       , txIn       :: ![TxIn]
-         -- | List of transaction outputs
-       , txOut      :: ![TxOut]
-         -- | The block number of timestamp at which this transaction is locked
-       , txLockTime :: !Word32
-       } deriving (Eq, Show, Read)
-
-instance NFData Tx where
-    rnf (Tx v i o l) = rnf v `seq` rnf i `seq` rnf o `seq` rnf l
-
-instance Binary Tx where
-
-    get = Tx <$> getWord32le
-             <*> (replicateList =<< get)
-             <*> (replicateList =<< get)
-             <*> getWord32le
-      where 
-        replicateList (VarInt c) = replicateM (fromIntegral c) get
-
-    put (Tx v is os l) = do
-        putWord32le v
-        put $ VarInt $ fromIntegral $ length is
-        forM_ is put
-        put $ VarInt $ fromIntegral $ length os
-        forM_ os put
-        putWord32le l
-
-instance FromJSON Tx where
-    parseJSON = withText "transaction" $ \t -> either fail return $
-        maybeToEither "tx not hex" (hexToBS $ T.unpack t) >>= decodeToEither
-
-instance ToJSON Tx where
-    toJSON = String . T.pack . bsToHex . encode'
-
--- | Data type representing the coinbase transaction of a 'Block'. Coinbase
--- transactions are special types of transactions which are created by miners
--- when they find a new block. Coinbase transactions have no inputs. They have
--- outputs sending the newly generated bitcoins together with all the block's
--- fees to a bitcoin address (usually the miners address). Data can be embedded
--- in a Coinbase transaction which can be chosen by the miner of a block. This
--- data also typically contains some randomness which is used, together with
--- the nonce, to find a partial hash collision on the block's hash.
-data CoinbaseTx = 
-    CoinbaseTx { 
-                 -- | Transaction data format version.
-                 cbVersion    :: !Word32
-                 -- | Previous outpoint. This is ignored for
-                 -- coinbase transactions but preserved for computing
-                 -- the correct txid.
-               , cbPrevOutput :: !OutPoint
-                 -- | Data embedded inside the coinbase transaction.
-               , cbData       :: !BS.ByteString
-                 -- | Transaction sequence number. This is ignored for
-                 -- coinbase transactions but preserved for computing
-                 -- the correct txid.
-               , cbInSequence :: !Word32
-                 -- | List of transaction outputs.
-               , cbOut        :: ![TxOut]
-                 -- | The block number of timestamp at which this 
-                 -- transaction is locked.
-               , cbLockTime   :: !Word32
-               } deriving (Eq, Show, Read)
-
-instance NFData CoinbaseTx where
-    rnf (CoinbaseTx v p d i o l) =
-        rnf v `seq` rnf p `seq` rnf d `seq` rnf i `seq` rnf o `seq` rnf l
-
-instance Binary CoinbaseTx where
-
-    get = do
-        v <- getWord32le
-        (VarInt len) <- get
-        unless (len == 1) $ fail "CoinbaseTx get: Input size is not 1"
-        op <- get
-        (VarInt cbLen) <- get
-        cb <- getByteString (fromIntegral cbLen)
-        sq <- getWord32le
-        (VarInt oLen) <- get
-        os <- replicateM (fromIntegral oLen) get
-        lt <- getWord32le
-        return $ CoinbaseTx v op cb sq os lt
-
-    put (CoinbaseTx v op cb sq os lt) = do
-        putWord32le v
-        put $ VarInt 1
-        put op
-        put $ VarInt $ fromIntegral $ BS.length cb
-        putByteString cb 
-        putWord32le sq
-        put $ VarInt $ fromIntegral $ length os
-        forM_ os put
-        putWord32le lt
-
--- | Data type representing a transaction input.
-data TxIn = 
-    TxIn { 
-           -- | Reference the previous transaction output (hash + position)
-           prevOutput   :: !OutPoint
-           -- | Script providing the requirements of the previous transaction
-           -- output to spend those coins.
-         , scriptInput  :: !BS.ByteString
-           -- | Transaction version as defined by the sender of the
-           -- transaction. The intended use is for replacing transactions with
-           -- new information before the transaction is included in a block.
-         , txInSequence :: !Word32
-         } deriving (Eq, Show, Read)
-
-instance NFData TxIn where
-    rnf (TxIn p i s) = rnf p `seq` rnf i `seq` rnf s
-
-instance Binary TxIn where
-
-    get = 
-        TxIn <$> get <*> (readBS =<< get) <*> getWord32le
-      where
-        readBS (VarInt len) = getByteString $ fromIntegral len
-
-    put (TxIn o s q) = do
-        put o 
-        put $ VarInt $ fromIntegral $ BS.length s
-        putByteString s
-        putWord32le q
-
--- | Data type representing a transaction output.
-data TxOut = 
-    TxOut { 
-            -- | Transaction output value.
-            outValue     :: !Word64
-            -- | Script specifying the conditions to spend this output.
-          , scriptOutput :: !BS.ByteString
-          } deriving (Eq, Show, Read)
-
-instance NFData TxOut where
-    rnf (TxOut v o) = rnf v `seq` rnf o
-
-instance Binary TxOut where
-
-    get = do
-        val <- getWord64le
-        unless (val <= 2100000000000000) $ fail $
-            "Invalid TxOut value: " ++ (show val)
-        (VarInt len) <- get
-        TxOut val <$> (getByteString $ fromIntegral len)
-
-    put (TxOut o s) = do
-        putWord64le o 
-        put $ VarInt $ fromIntegral $ BS.length s
-        putByteString s
-
--- | The OutPoint is used inside a transaction input to reference the previous
--- transaction output that it is spending.
-data OutPoint = 
-    OutPoint { 
-               -- | The hash of the referenced transaction.
-               outPointHash  :: !TxHash
-               -- | The position of the specific output in the transaction.
-               -- The first output position is 0.
-             , outPointIndex :: !Word32
-             } deriving (Read, Show, Eq)
-
-instance NFData OutPoint where
-    rnf (OutPoint h i) = rnf h `seq` rnf i
-
-instance FromJSON OutPoint where
-    parseJSON = withText "outpoint" $ \t -> either fail return $
-        maybeToEither "outpoint not hex" 
-          (hexToBS $ T.unpack t) >>= decodeToEither
-
-instance ToJSON OutPoint where
-    toJSON = String . T.pack . bsToHex . encode'
-
-instance Binary OutPoint where
-    get = do
-        (h,i) <- liftM2 (,) get getWord32le
-        return $ OutPoint h i
-    put (OutPoint h i) = put h >> putWord32le i
 
 -- | Data type representing a variable length integer. The 'VarInt' type
 -- usually precedes an array or a string that can vary in length.
