+++ Network/Haskoin/Script/SigHash.hs
@@ -36,7 +36,8 @@ import Network.Haskoin.Crypto.BigWord
 import Network.Haskoin.Crypto.Hash
 import Network.Haskoin.Crypto.ECDSA
 import Network.Haskoin.Script.Types
-import Network.Haskoin.Protocol
+import Network.Haskoin.Protocol.Types
+import Network.Haskoin.Transaction.Types
 import Network.Haskoin.Util
 
 -- | Data type representing the different ways a transaction can be signed.
