+++ Network/Haskoin/Transaction/Builder.hs
@@ -41,6 +41,7 @@ import Network.Haskoin.Util
 import Network.Haskoin.Crypto
 import Network.Haskoin.Protocol
 import Network.Haskoin.Script
+import Network.Haskoin.Transaction.Types
 
 -- | A Coin is an output of a transaction that can be spent by another
 -- transaction.
