+++ haskoin.cabal
@@ -36,6 +36,7 @@ library
                        Network.Haskoin.Protocol,
                        Network.Haskoin.Script,
                        Network.Haskoin.Transaction,
+                       Network.Haskoin.Block,
                        Network.Haskoin.Internals,
                        Network.Haskoin.Constants
     other-modules:     Network.Haskoin.Crypto.NumberTheory, 
@@ -57,6 +58,7 @@ library
                        Network.Haskoin.Script.Parser, 
                        Network.Haskoin.Script.SigHash,
                        Network.Haskoin.Script.Evaluator,
+                       Network.Haskoin.Transaction.Types,
                        Network.Haskoin.Transaction.Builder
     build-depends:     aeson >= 0.8 && < 0.9,
                        base >= 4.6 && < 5, 
@@ -103,14 +105,16 @@ test-suite test-haskoin
                        Network.Haskoin.Crypto.Arbitrary,
                        Network.Haskoin.Crypto.Units,
                        Network.Haskoin.Protocol.Arbitrary,
+                       Network.Haskoin.Protocol.ArbitraryMsg,
                        Network.Haskoin.Protocol.Tests,
-                       Network.Haskoin.Protocol.Units,
                        Network.Haskoin.Script.Arbitrary,
                        Network.Haskoin.Script.Tests,
                        Network.Haskoin.Script.Units,
                        Network.Haskoin.Transaction.Tests,
                        Network.Haskoin.Transaction.Units,
                        Network.Haskoin.Transaction.Arbitrary,
+                       Network.Haskoin.Block.Tests,
+                       Network.Haskoin.Block.Arbitrary,
                        Network.Haskoin.Stratum.Tests,
                        Network.Haskoin.Stratum.Units,
                        Network.Haskoin.Stratum.Arbitrary,
