+++ tests/Network/Haskoin/Protocol/Arbitrary.hs
@@ -5,7 +5,6 @@
 module Network.Haskoin.Protocol.Arbitrary () where
 
 import Test.QuickCheck
-import Network.Haskoin.Crypto.Arbitrary()
 
 import Control.Monad
 import Control.Applicative 
@@ -14,6 +13,7 @@ import qualified Data.Sequence as S (fromList)
 
 import Network.Socket
 
+import Network.Haskoin.Crypto.Arbitrary()
 import Network.Haskoin.Protocol
 import Network.Haskoin.Crypto
 
@@ -76,80 +76,8 @@ instance Arbitrary RejectCode where
                          , RejectCheckpoint
                          ]
 
-instance Arbitrary BlockHeader where
-    arbitrary = BlockHeader <$> arbitrary
-                            <*> (fromIntegral . hash256 <$> arbitrary)
-                            <*> (hash256 <$> arbitrary)
-                            <*> arbitrary
-                            <*> arbitrary
-                            <*> arbitrary
-                            
-instance Arbitrary Tx where
-    arbitrary = do
-        v   <- arbitrary
-        tin <- do 
-            l <- choose (0,10)
-            vectorOf l arbitrary
-        tout <- do
-            l <- choose (0,10)
-            vectorOf l arbitrary
-        t    <- arbitrary
-        return $ Tx v tin tout t
-
-instance Arbitrary CoinbaseTx where
-    arbitrary = CoinbaseTx <$> arbitrary
-                           <*> (return $ OutPoint 0 0xffffffff)
-                           <*> arbitrary
-                           <*> (return $ 0xffffffff)
-                           <*> (listOf arbitrary)
-                           <*> arbitrary
-
-instance Arbitrary TxIn where
-    arbitrary = TxIn <$> arbitrary
-                     <*> arbitrary
-                     <*> arbitrary
-
-instance Arbitrary TxOut where
-    arbitrary = TxOut <$> (choose (0,2100000000000000))
-                      <*> arbitrary
-
-instance Arbitrary OutPoint where
-    arbitrary = OutPoint <$> arbitrary
-                         <*> arbitrary
-
-instance Arbitrary Block where
-    arbitrary = do
-        h <- arbitrary
-        c <- arbitrary
-        t <- do 
-            l <- choose (0,10)
-            vectorOf l arbitrary
-        return $ Block h c t
-
-instance Arbitrary MerkleBlock where
-    arbitrary = do
-        h <- arbitrary
-        ntx <- arbitrary
-        hashes <- arbitrary
-        c <- choose (1,10)
-        flags <- vectorOf (c*8) arbitrary
-        return $ MerkleBlock h ntx hashes flags
-
-instance Arbitrary GetBlocks where
-    arbitrary = GetBlocks <$> arbitrary
-                          <*> (listOf arbitrary)
-                          <*> arbitrary
-
 instance Arbitrary GetData where
     arbitrary = GetData <$> (listOf arbitrary)
-
-instance Arbitrary GetHeaders where
-    arbitrary = GetHeaders <$> arbitrary
-                           <*> (listOf arbitrary)
-                           <*> arbitrary
-
-instance Arbitrary Headers where
-    arbitrary = Headers <$> (listOf (liftM2 (,) arbitrary arbitrary))
 
 instance Arbitrary NotFound where
     arbitrary = NotFound <$> (listOf arbitrary)
@@ -181,34 +109,6 @@ instance Arbitrary MessageCommand where
                          , MCPong
                          , MCAlert
                          ]
-
-instance Arbitrary MessageHeader where
-    arbitrary = MessageHeader <$> arbitrary
-                              <*> arbitrary
-                              <*> arbitrary
-                              <*> arbitrary
-
-instance Arbitrary Message where
-    arbitrary = oneof [ MVersion    <$> arbitrary
-                      , return MVerAck
-                      , MAddr        <$> arbitrary
-                      , MInv         <$> arbitrary
-                      , MGetData     <$> arbitrary
-                      , MNotFound    <$> arbitrary
-                      , MGetBlocks   <$> arbitrary
-                      , MGetHeaders  <$> arbitrary
-                      , MTx          <$> arbitrary
-                      , MBlock       <$> arbitrary
-                      , MMerkleBlock <$> arbitrary
-                      , MHeaders     <$> arbitrary
-                      , return MGetAddr
-                      , MFilterLoad  <$> arbitrary
-                      , MFilterAdd   <$> arbitrary
-                      , return MFilterClear
-                      , MPing        <$> arbitrary
-                      , MPong        <$> arbitrary
-                      , MAlert       <$> arbitrary
-                      ]
 
 instance Arbitrary BloomFlags where
     arbitrary = elements [ BloomUpdateNone
